src/consumer/requirements.txt
pandas
kafka-python